+++ src/aux-common/bots/BotEvents.ts
@@ -3302,13 +3302,47 @@ export interface DefineGlobalBotAction extends AsyncAction {
 export const APPROVED_SYMBOL = Symbol('approved');
 
 /**
- * Defines an interface that represents the base for actions that deal with records.
+ * Defines an interface that represents the base for options for a records action.
  */
-export interface RecordsAction extends AsyncAction {
+export interface RecordActionOptions {
     /**
      * The HTTP endpoint that the request should interface with.
      */
     endpoint?: string;
+}
+
+/**
+ * Defines an interface that represents the base for actions that deal with records.
+ */
+export interface RecordsAction extends AsyncAction {
+    /**
+     * The options that the action should use.
+     */
+    options: RecordActionOptions;
+}
+
+/**
+ * Defines a type that represents a policy that indicates which users are allowed to affect a record.
+ * 
+ * True indicates that any user can edit the record.
+ * An array of strings indicates the list of users that are allowed to edit the record.
+ */
+export type RecordUserPolicyType = true | string[];
+
+/**
+ * The options for data record actions.
+ */
+export interface DataRecordOptions extends RecordActionOptions {
+
+    /**
+     * The policy that should be used for updating the record.
+     */
+    updatePolicy?: RecordUserPolicyType;
+
+    /**
+     * The policy that should be used for deleting the record.
+     */
+    deletePolicy?: RecordUserPolicyType;
 }
 
 /**
@@ -3349,6 +3383,8 @@ export interface RecordDataAction extends DataRecordAction {
      * The data that should be recorded.
      */
     data: any;
+
+    options: DataRecordOptions;
 }
 
 /**
@@ -6302,7 +6338,7 @@ export function getPublicRecordKey(
  * @param address The address that the data should be stored at in the record.
  * @param data The data to store.
  * @param requiresApproval Whether to try to record data that requires approval.
- * @param endpoint The HTTP Origin that should be used for the records request.
+ * @param options The options that should be used for the action.
  * @param taskId The ID of the task.
  */
 export function recordData(
@@ -6310,7 +6346,7 @@ export function recordData(
     address: string,
     data: any,
     requiresApproval: boolean,
-    endpoint: string,
+    options: DataRecordOptions,
     taskId: number | string
 ): RecordDataAction {
     return {
@@ -6319,7 +6355,7 @@ export function recordData(
         address,
         data,
         requiresApproval,
-        endpoint,
+        options,
         taskId,
     };
 }
@@ -6329,14 +6365,14 @@ export function recordData(
  * @param recordName The name of the record to retrieve.
  * @param address The address of the data to retrieve.
  * @param requiresApproval Whether to try to get a record that requires manual approval.
- * @param endpoint The HTTP Origin that should be used for the records request.
+ * @param options The options that should be used for the action.
  * @param taskId The ID of the task.
  */
 export function getRecordData(
     recordName: string,
     address: string,
     requiresApproval: boolean,
-    endpoint: string,
+    options: RecordActionOptions,
     taskId?: number | string
 ): GetRecordDataAction {
     return {
@@ -6344,7 +6380,7 @@ export function getRecordData(
         recordName,
         address,
         requiresApproval,
-        endpoint,
+        options,
         taskId,
     };
 }
@@ -6353,13 +6389,13 @@ export function getRecordData(
  * Creates a ListRecordDataAction.
  * @param recordName The name of the record.
  * @param startingAddress The address that the list should start with.
- * @param endpoint The HTTP Origin that should be used for the records request.
+ * @param options The options that should be used for the action.
  * @param taskId The ID of the task.
  */
 export function listDataRecord(
     recordName: string,
     startingAddress: string,
-    endpoint: string,
+    options: RecordActionOptions,
     taskId?: number | string
 ): ListRecordDataAction {
     return {
@@ -6367,7 +6403,7 @@ export function listDataRecord(
         recordName,
         startingAddress,
         requiresApproval: false,
-        endpoint,
+        options,
         taskId,
     };
 }
@@ -6377,14 +6413,14 @@ export function listDataRecord(
  * @param recordKey The key that should be used to access the record.
  * @param address The address of the data to erase.
  * @param requiresApproval Whether to try to erase a record that requires manual approval.
- * @param endpoint The HTTP Origin that should be used for the records request.
+ * @param options The options that should be used for the action.
  * @param taskId The ID of the task.
  */
 export function eraseRecordData(
     recordKey: string,
     address: string,
     requiresApproval: boolean,
-    endpoint: string,
+    options: RecordActionOptions,
     taskId?: number | string
 ): EraseRecordDataAction {
     return {
@@ -6392,7 +6428,7 @@ export function eraseRecordData(
         recordKey,
         address,
         requiresApproval,
-        endpoint,
+        options,
         taskId,
     };
 }
@@ -6414,14 +6450,14 @@ export function approveDataRecord<T extends DataRecordAction>(action: T): T {
  * @param data The data to store.
  * @param description The description of the file.
  * @param mimeType The MIME type of the file.
- * @param endpoint The HTTP Origin that should be used for the records request.
+ * @param options The options that should be used for the action.
  */
 export function recordFile(
     recordKey: string,
     data: any,
     description: string,
     mimeType: string,
-    endpoint: string,
+    options: RecordActionOptions,
     taskId?: number | string
 ): RecordFileAction {
     return {
@@ -6430,7 +6466,7 @@ export function recordFile(
         data,
         description,
         mimeType,
-        endpoint,
+        options,
         taskId,
     };
 }
@@ -6439,20 +6475,20 @@ export function recordFile(
  * Creates a EraseFileAction.
  * @param recordKey The key that should be used to access the record.
  * @param fileUrl The URL that the file was stored at.
- * @param endpoint The HTTP Origin that should be used for the records request.
+ * @param options The options that should be used for the action.
  * @param taskId The ID of the task.
  */
 export function eraseFile(
     recordKey: string,
     fileUrl: string,
-    endpoint: string,
+    options: RecordActionOptions,
     taskId?: number | string
 ): EraseFileAction {
     return {
         type: 'erase_file',
         recordKey,
         fileUrl,
-        endpoint,
+        options,
         taskId,
     };
 }
@@ -6462,14 +6498,14 @@ export function eraseFile(
  * @param recordKey The key that should be used to access the record.
  * @param eventName The name of the event.
  * @param count The number of times that the event occurred.
- * @param endpoint The HTTP Origin that should be used for the records request.
+ * @param options The options that should be used for the action.
  * @param taskId The Id of the task.
  */
 export function recordEvent(
     recordKey: string,
     eventName: string,
     count: number,
-    endpoint: string,
+    options: RecordActionOptions,
     taskId?: number | string
 ): RecordEventAction {
     return {
@@ -6477,7 +6513,7 @@ export function recordEvent(
         recordKey,
         eventName,
         count,
-        endpoint,
+        options,
         taskId,
     };
 }
@@ -6486,20 +6522,20 @@ export function recordEvent(
  * Creates a GetEventCountAction.
  * @param recordName The name of the record.
  * @param eventName The name of the events.
- * @param endpoint The HTTP Origin that should be used for the records request.
+ * @param options The options that should be used for the action.
  * @param taskId The ID.
  */
 export function getEventCount(
     recordName: string,
     eventName: string,
-    endpoint: string,
+    options: RecordActionOptions,
     taskId?: number | string
 ): GetEventCountAction {
     return {
         type: 'get_event_count',
         recordName,
         eventName,
-        endpoint,
+        options,
         taskId,
     };
 }