+++ src/aux-common/runtime/AuxLibrary.spec.ts
@@ -4613,7 +4613,7 @@ describe('AuxLibrary', () => {
                     'address',
                     { data: true },
                     false,
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4632,7 +4632,31 @@ describe('AuxLibrary', () => {
                     'address',
                     { data: true },
                     false,
-                    'myEndpoint',
+                    {
+                        endpoint: 'myEndpoint'
+                    },
+                    context.tasks.size
+                );
+                expect(action[ORIGINAL_OBJECT]).toEqual(expected);
+                expect(context.actions).toEqual([expected]);
+            });
+
+            it('should support custom options', async () => {
+                const action: any = library.api.os.recordData(
+                    'recordKey',
+                    'address',
+                    { data: true },
+                    { updatePolicy: true, endpoint: 'myEndpoint' }
+                );
+                const expected = recordData(
+                    'recordKey',
+                    'address',
+                    { data: true },
+                    false,
+                    {
+                        updatePolicy: true,
+                        endpoint: 'myEndpoint'
+                    },
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4655,7 +4679,7 @@ describe('AuxLibrary', () => {
                         },
                     },
                     false,
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4680,7 +4704,7 @@ describe('AuxLibrary', () => {
                         },
                     },
                     false,
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4700,7 +4724,7 @@ describe('AuxLibrary', () => {
                     'address',
                     { data: true },
                     true,
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4719,7 +4743,26 @@ describe('AuxLibrary', () => {
                     'address',
                     { data: true },
                     true,
-                    'myEndpoint',
+                    { endpoint: 'myEndpoint' },
+                    context.tasks.size
+                );
+                expect(action[ORIGINAL_OBJECT]).toEqual(expected);
+                expect(context.actions).toEqual([expected]);
+            });
+
+            it('should support custom options', async () => {
+                const action: any = library.api.os.recordManualApprovalData(
+                    'recordKey',
+                    'address',
+                    { data: true },
+                    { updatePolicy: true, endpoint: 'myEndpoint'}
+                );
+                const expected = recordData(
+                    'recordKey',
+                    'address',
+                    { data: true },
+                    true,
+                    { updatePolicy: true, endpoint: 'myEndpoint' },
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4742,7 +4785,7 @@ describe('AuxLibrary', () => {
                         },
                     },
                     true,
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4767,7 +4810,7 @@ describe('AuxLibrary', () => {
                         },
                     },
                     true,
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4785,7 +4828,7 @@ describe('AuxLibrary', () => {
                     'recordKey',
                     'address',
                     false,
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4802,7 +4845,7 @@ describe('AuxLibrary', () => {
                     'recordKey',
                     'address',
                     false,
-                    'myEndpoint',
+                    { endpoint: 'myEndpoint'},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4818,7 +4861,7 @@ describe('AuxLibrary', () => {
                     'recordName',
                     'address',
                     false,
-                    null,
+                    { },
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4834,7 +4877,7 @@ describe('AuxLibrary', () => {
                     'recordName',
                     'address',
                     false,
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4852,7 +4895,7 @@ describe('AuxLibrary', () => {
                     'recordKey',
                     'address',
                     true,
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4869,7 +4912,7 @@ describe('AuxLibrary', () => {
                     'recordKey',
                     'address',
                     true,
-                    'myEndpoint',
+                    { endpoint: 'myEndpoint' },
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4885,7 +4928,7 @@ describe('AuxLibrary', () => {
                     'recordName',
                     'address',
                     true,
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4901,7 +4944,7 @@ describe('AuxLibrary', () => {
                     'recordName',
                     'address',
                     true,
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4915,7 +4958,7 @@ describe('AuxLibrary', () => {
                 const expected = listDataRecord(
                     'recordName',
                     null,
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4931,7 +4974,7 @@ describe('AuxLibrary', () => {
                 const expected = listDataRecord(
                     'recordName',
                     null,
-                    'myEndpoint',
+                    { endpoint: 'myEndpoint' },
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4946,7 +4989,7 @@ describe('AuxLibrary', () => {
                 const expected = listDataRecord(
                     'recordName',
                     'address',
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4961,7 +5004,7 @@ describe('AuxLibrary', () => {
                 const expected = listDataRecord(
                     'recordName',
                     'address',
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4976,7 +5019,7 @@ describe('AuxLibrary', () => {
                 const expected = listDataRecord(
                     'recordName',
                     'address',
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4994,7 +5037,7 @@ describe('AuxLibrary', () => {
                     'recordKey',
                     'address',
                     false,
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -5011,7 +5054,7 @@ describe('AuxLibrary', () => {
                     'recordKey',
                     'address',
                     false,
-                    'myEndpoint',
+                    { endpoint: 'myEndpoint' },
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -5053,7 +5096,7 @@ describe('AuxLibrary', () => {
                     'recordKey',
                     'address',
                     true,
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -5070,7 +5113,7 @@ describe('AuxLibrary', () => {
                     'recordKey',
                     'address',
                     true,
-                    'myEndpoint',
+                    { endpoint:'myEndpoint' },
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -5119,7 +5162,7 @@ describe('AuxLibrary', () => {
                     'data',
                     'description',
                     undefined,
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -5140,7 +5183,7 @@ describe('AuxLibrary', () => {
                     'data',
                     'description',
                     undefined,
-                    'https://localhost:5000',
+                    { endpoint: 'https://localhost:5000' },
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -5157,7 +5200,7 @@ describe('AuxLibrary', () => {
                     'data',
                     undefined,
                     undefined,
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -5177,7 +5220,7 @@ describe('AuxLibrary', () => {
                     'data',
                     undefined,
                     'image/png',
-                    null,
+                    { },
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -5218,7 +5261,7 @@ describe('AuxLibrary', () => {
                     },
                     undefined,
                     undefined,
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -5295,7 +5338,7 @@ describe('AuxLibrary', () => {
                 const expected = eraseFile(
                     'recordKey',
                     'fileUrl',
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -5311,7 +5354,7 @@ describe('AuxLibrary', () => {
                 const expected = eraseFile(
                     'recordKey',
                     'fileUrl',
-                    'http://localhost:5000',
+                    { endpoint: 'http://localhost:5000' },
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -5349,7 +5392,7 @@ describe('AuxLibrary', () => {
                     'recordKey',
                     'eventName',
                     1,
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -5366,7 +5409,7 @@ describe('AuxLibrary', () => {
                     'recordKey',
                     'eventName',
                     1,
-                    'http://localhost:5000',
+                    { endpoint: 'http://localhost:5000' },
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -5407,7 +5450,7 @@ describe('AuxLibrary', () => {
                 const expected = getEventCount(
                     'recordKey',
                     'eventName',
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -5423,7 +5466,7 @@ describe('AuxLibrary', () => {
                 const expected = getEventCount(
                     'recordKey',
                     'eventName',
-                    'http://localhost:5000',
+                    { endpoint: 'http://localhost:5000' },
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -5438,7 +5481,7 @@ describe('AuxLibrary', () => {
                 const expected = getEventCount(
                     'recordName',
                     'eventName',
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -5453,7 +5496,7 @@ describe('AuxLibrary', () => {
                 const expected = getEventCount(
                     'recordName',
                     'eventName',
-                    null,
+                    {},
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);